refactor(gen_reply): unify voice synthesis into the media generation module
Move the spoken-reply service alongside image, video, and music so all media
generators share one home and one Generator-style calling convention.
